convert all implicit integers (eg 1) in explicit (eg 1L))

+++ tests/testthat/test-utils.R
@@ -8,28 +8,28 @@ test_that("assert_is_single_h_desc works", {
   )
   expect_true(assert_is_h_desc(test_desc))
 
-  expect_true(assert_is_single_h_desc(test_desc[[1]]))
-  expect_true(assert_is_h_desc(test_desc[[1]]))
+  expect_true(assert_is_single_h_desc(test_desc[[1L]]))
+  expect_true(assert_is_h_desc(test_desc[[1L]]))
 })
 
 
 test_that("is_val_freq_list works", {
   expect_false(is_val_freq_list(integer()))
-  expect_false(is_val_freq_list(integer(1)))
-  expect_false(is_val_freq_list(list(a = 1, b = 2)))
+  expect_false(is_val_freq_list(integer(1L)))
+  expect_false(is_val_freq_list(list(a = 1L, b = 2L)))
 
   expect_true(
-    is_val_freq_list(list(value = 1, frequency = 2))
+    is_val_freq_list(list(value = 1L, frequency = 2L))
   )
 })
 
 
 test_that("is_proper_matrix works", {
-  expect_false(is_proper_matrix(1))
-  expect_false(is_proper_matrix(matrix(c(1, 2))))
-  expect_false(is_proper_matrix(matrix(c(1, 2), nrow = 1)))
+  expect_false(is_proper_matrix(1L))
+  expect_false(is_proper_matrix(matrix(c(1L, 2L))))
+  expect_false(is_proper_matrix(matrix(c(1L, 2L), nrow = 1L)))
 
-  expect_true(is_proper_matrix(matrix(1:4, nrow = 2)))
+  expect_true(is_proper_matrix(matrix(1L:4L, nrow = 2L)))
 })
 
 
@@ -48,5 +48,5 @@ test_that("fake_h_group_test works", {
     "Only one group"
   )
   expect_type(out, "list")
-  expect_equal(out[["P"]], c(P = 1))
+  expect_equal(out[["P"]], c(P = 1L))
 })
